backend/requirements.txt
# MedScribe Backend Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=12.0

# Database
sqlalchemy>=2.0.30
alembic>=1.13.0
aiosqlite>=0.20.0
asyncpg>=0.30.0

# Authentication & Security
pyjwt>=2.8.0
bcrypt>=4.1.0
cryptography>=42.0.0
passlib[bcrypt]>=1.7.4

# AI — Claude API
anthropic>=0.25.0

# Clinical NLP — medspaCy for medical entity extraction, negation detection, section detection
# Install spaCy model separately: python -m spacy download en_core_web_sm
medspacy>=1.3.0
spacy>=3.7.0

# PDF Generation
reportlab>=4.2.0

# HTTP Client
httpx>=0.27.0

# Configuration
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Utilities
python-dateutil>=2.9.0

# Rate Limiting
slowapi>=0.1.9

# Validation
email-validator>=2.1.0
